modifs dans le stockage du graphe et bon temps pour ALT

+++ Code/Algorithme.h
@@ -9,9 +9,9 @@ protected:
     vector<Arc>* A;
     double timer;
     double duration;
-    long visites;
-    Chemin reconstitution_chemin_forward(long s, long t, vector<long> *distance) const;
-    Chemin reconstitution_chemin_backward(long s, long t, vector<long> *distance) const;
+    int visites;
+    Chemin reconstitution_chemin_forward(int s, int t, vector<long> *distance) const;
+    Chemin reconstitution_chemin_backward(int s, int t, vector<long> *distance) const;
 
 public:
     Algorithme(Graphe* g);
@@ -21,14 +21,14 @@ public:
     double now();
     double get_duration(){return duration;}
     void add_visite();
-    long get_visites(){return visites;}
+    int get_visites(){return visites;}
 
     // REQUETES (les variantes bidirectionnelles ne sont pas partout implémentées
-    virtual long requete(long s, long t, bool verbose = false)=0;
-    virtual long requete_bi(long s, long t, bool verbose = false)=0;
+    virtual long requete(int s, int t, bool verbose = false)=0;
+    virtual long requete_bi(int s, int t, bool verbose = false)=0;
 
-    virtual pair<long, Chemin> chemin(long s, long t, bool verbose = false)=0;
-    virtual pair<long, Chemin> chemin_bi(long s, long t, bool verbose = false)=0;
+    virtual pair<long, Chemin> chemin(int s, int t, bool verbose = false)=0;
+    virtual pair<long, Chemin> chemin_bi(int s, int t, bool verbose = false)=0;
 
     vector<Sommet>* get_V(){return V;}
     vector<Arc>* get_A(){return A;}
