+++ Code/Chemin.h
@@ -8,12 +8,12 @@
 using namespace std;
 
 class Chemin{
-    vector<long> listeId;  //liste des indices des sommets formant le chemin
+    vector<int> listeId;  //liste des indices des sommets formant le chemin
 public:
-    void push_back(long id);
+    void push_back(int id);
     void inversion();   //inverser le sens du chemin
     Chemin join(Chemin C);
     string chaine() const;
-    double longueur(vector<Arc> *A, vector<Sommet>* V) const;
+    long longueur(vector<Arc> *A, vector<Sommet>* V) const;
 };
 ostream& operator<<(ostream&str,const Chemin &C);
