+++ Code/Dijkstra.h
@@ -11,7 +11,5 @@ class Dijkstra: public AStar{
 public:
     Dijkstra (Graphe* g):AStar(g){}
     ~Dijkstra(){}
-    virtual long pi(long u, long t, long s){return 0;}
+    virtual long pi(int u, int t, int s){return 0;}
 };
-
-
